requirements.txt
aniso8601==9.0.1
arabic-reshaper==3.0.0
arrow==1.2.2
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.0.0
charset-normalizer==3.3.2
click==8.1.3
docopt==0.6.2
filelock==3.13.1
Flask==2.2.2
Flask-Cors==4.0.0
fsspec==2023.10.0
gunicorn==20.1.0
huggingface-hub==0.19.4
idna==3.4
imageio==2.33.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.3.2
lazy_loader==0.3
lxml==5.1.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
opencv-python==4.9.0.80
packaging==23.2
Pillow==9.4.0
pipreqs==0.4.13
python-dateutil==2.8.2
python-slugify==6.1.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safe-gpu==1.5.1
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.2
six==1.16.0
sympy==1.12
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2023.12.9
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.8.1
